integrated with spreadsheet
