+++ index.js
@@ -1,15 +1,15 @@
 require('dotenv').config();
 const parseArgs = require('minimist');
-const rimraf = require('rimraf');
 
 const runEnvChecks = require('./src/envChecks');
 const runGeneralChecks = require('./src/generalChecks');
 const runGitChecks = require('./src/gitChecks');
 const runSeoChecks = require('./src/seoChecks');
-const runEslintChecks = require('./src/linterChecks');
-const { green } = require('./src/constants/colors');
+const runBuildChecks = require('./src/seoChecks');
 
-const shell = require('shelljs');
+const { green } = require('./src/constants/colors');
+const writeSpreadSheet = require('./src/utils/writeSheet');
+
 const techs = {
   react: require('./src/reactChecks'),
   angular: require('./src/angularChecks'),
@@ -25,8 +25,6 @@ const csvWriter = createCsvWriter({
     { id: 'value', title: 'RESULTADO' }
   ]
 });
-
-shell.config.silent = true;
 
 const args = parseArgs(process.argv);
 
@@ -67,28 +65,12 @@ if (args.org) {
   organization = args.o;
 }
 
-function runBuild() {
-  const seconds = 1000;
-  console.log(green, 'Empezando instalacion de dependencias para el build...');
-  shell.exec(`npm i --prefix ./${testPath}`);
-  const eslintData = runEslintChecks(testPath);
-  console.log(green, 'Chequeos de eslint terminados con exito ✓');
-  console.log(green, 'Generando el build...');
-  const start = new Date();
-  shell.exec(`npm run build development --prefix ./${testPath}`);
-  const buildTime = (new Date().getTime() - start.getTime()) / seconds;
-  console.log(green, 'Build terminado con exito ✓');
-  rimraf.sync(`./${testPath}/node_modules`);
-  rimraf.sync(`./${testPath}/build`);
-  return [...eslintData, { metric: 'Build', description: 'Build Time', value: `${buildTime}` }];
-}
-
 async function executeChecks() {
   let gitData = [];
   let envData = [];
   let generalData = [];
   let techData = [];
-  let buildData = [];
+  // let buildData = [];
   const eslintData = [];
   if (args.onlyGit) {
     gitData = await runGitChecks(repoName, organization);
@@ -102,13 +84,14 @@ async function executeChecks() {
     console.log(green, 'Chequeos de github terminados con exito ✓');
     techData = await techs[techChecks](testPath);
     console.log(green, 'Chequeos de tecnologia terminados con exito ✓');
-    buildData = await runBuild(testPath);
+    // buildData = runBuild(testPath);
   }
-  return [...envData, ...generalData, ...gitData, ...techData, ...buildData, ...eslintData];
+  return [...envData, ...generalData, ...gitData, ...techData, ...eslintData];
 }
 
 async function executeAudit() {
   const reports = await executeChecks();
+  writeSpreadSheet(reports, testPath);
   csvWriter.writeRecords(reports);
   console.log(green, 'Chequeos terminados con exito ✓');
 }
